+++ restapi/server.ts
@@ -7,6 +7,15 @@ import api from "./api";
 const app: Application = express();
 const port: number = 5000;
 
+const { MongoClient, ServerApiVersion } = require('mongodb');
+const uri = "mongodb+srv://dede_5c:cuentademongo@dede.irypt.mongodb.net/DeDe?retryWrites=true&w=majority";
+const client = new MongoClient(uri, { useNewUrlParser: true, useUnifiedTopology: true, serverApi: ServerApiVersion.v1 });
+client.connect(err => {
+  const collection = client.db("test").collection("devices");
+  // perform actions on the collection object
+  client.close();
+});
+
 const options: cors.CorsOptions = {
   origin: ['http://localhost:3000']
 };
